requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
networkx==3.2.1
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.0
httpx==0.25.0
